Move all rendering to single module
